backend/requirements.txt
# backend dependencies

flask
flask-cors
langchain
langchain_openai
langchain_community
langgraph
python-dotenv
openai
tavily-python